pyproject.toml
[project]
name = "workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pygame-menu==4.4.8",
    "pygame-widgets==1.1.5",
    "pygameshader==1.0.10",
]
